# apis4lawenforcement_swaggerdef.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "APIs 4 SAPOL Demo"
  description: SAPol provides a simple API allowing consumers to view people, license, vehicle and property info.
  #license:
  #  name: MIT
  #  url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
...
basePath: "/"
host: "localhost:3000"
...
schemes:
- http
security:
  - X-MyAPI-Key: []
consumes:
- application/json
produces:
- application/json
paths:
...
  /poi:
    x-summary: Persons of Interest 
    post:
      tags:
      - "Persons Of Interest"
      description: 'Create a Person of Interest record.'
      summary: 'Register a Person of Interest / Contact Information' 
      parameters:  
        - name: body
          in: body
          required: true
          schema:
            properties:
              PersonOfInterest:
                $ref: '#/definitions/PersonsOfInterestRequest'
      responses:
        201:
          description: 'Successfully created Person Of Interest record. It is assumed that the database will auto-generate the ID'
          schema:
            properties:
              PersonOfInterest:
                $ref: '#/definitions/PersonsOfInterestResponse'
          examples:
            application/json:
              PersonOfInterest:
                id: 1
                fname: 'Dave'
                lname: 'Redbull'
                dob: '01-JAN-1980'
                address: '20 Charlesworthnut Street, Walkerville, SA 5000'
                facialImageURL: '/foo/image/123.jpg'
                licenseNumber: 'A12345'
                criminalHistorySummary: 'hardened criminal'
                hasWarrantsOutstanding: 'YES'
                hasFinesOutstanding: 'YES'
              
    get:
      tags:
      - "Persons Of Interest"
      summary: "List containing Persons of Interest Details filtered by parameters provided"
      description: "List containing Persons of Interest Details filtered by parameters provided"
      operationId: "getPersonsOfInterest"
      produces:
        - "application/json"
      parameters:
      - name: "firstName"
        in: "query"
        description: "firstName of the contact used to filter the record(s) to be returned."
        required: false
        type: "string"
        x-example: 'Dave'
      - name: "lastName"
        in: "query"
        description: "lastName of the contact used to filter the record(s) to be returned."
        required: false
        type: "string"
        x-example: 'Redbull'
      - name: "dateOfBirth"
        in: "query"
        description: "date of birth of the contact used to filter the record(s) to be returned."
        required: false
        type: "string"
        x-example: '01-JAN-1980'
      responses:
        200:
          description: Successful Task Lookup Response
          schema:
            properties:
              PersonsOfInterest:
                type: array
                items:
                  $ref: '#/definitions/PersonsOfInterestResponse'
          examples:
            application/json:
              PersonsOfInterest:
              - id: 1
                fname: 'Dave'
                lname: 'Redbull'
                dob: '01-JAN-1980'
                address: '20 Charlesworthnut Street, Walkerville, SA 5000'
                facialImageURL: '/foo/image/123.jpg'
                licenseNumber: 'A12345'
                criminalHistorySummary: 'x'
                hasWarrantsOutstanding: 'YES'
                hasFinesOutstanding: 'YES'
              - id: 2
                fname: 'Daniel'
                lname: 'Truedea'
                dob: '01-JAN-1980'
                address: '1 Hindley Street, Adelaide, SA 5000'
                facialImageURL: '/foo/image/243.jpg'
                licenseNumber: 'A54632'
                criminalHistorySummary: 'Handle with extreme care. Prone to verbal spray, spitting and violent outbursts.'
                hasWarrantsOutstanding: 'NO'
                hasFinesOutstanding: 'YES'
  /poi/{personId}:
    put:
      tags:
      - "Persons Of Interest"
      summary: 'Update the details of a Person of Interest'
      description: 'Update a contact within the Police Citizen database.'
      operationId: "updatePersonOfInterestById"
      produces:
      - "application/json"
      parameters:
        - name: 'personId'
          in: path
          required: true
          type: number
          x-example: 1
        - name: body
          in: body
          required: true
          schema:
            properties:
              PersonOfInterest:
                $ref: '#/definitions/PersonsOfInterestRequest'
      responses:
        201:
          description: 'Successful update of Persons Of Interest information'
          schema:
            $ref: '#/definitions/PersonsOfInterestResponse'
          examples:
            application/json:
              PersonOfInterest:
                id: 1
                fname: 'Dave'
                lname: 'Redbull'
                dob: '01-JAN-1980'
                address: '20 Charlesworthnut Street, Walkerville, SA 5000'
                facialImageURL: '/foo/image/123.jpg'
                licenseNumber: 'A12345'
                license_class: 'B'
                licenseExpiryDate: '02-MAR-2020'
                criminalHistorySummary: 'Dave was named a person of interest by the coroner re case ACC555'
                hasWarrantsOutstanding: ''
                hasFinesOutstanding: 'Yes'
...
securityDefinitions:
  X-MyAPI-Key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
definitions:
...
  PersonsOfInterestRequest:
    title: "Persons Of Interest Request"
    type: object
    required:
    - "fname"
    - "lname"
    - "address"
    - "dob"
    properties:
      fname:
        type: string
        example: 'Dave'
      lname:
        type: string
        example: 'Dundocerty'
      dob:
        type: string
        example: '19-MAR-2000'
      address:
        type: string
        example: '25 Grenfell Street, Adelaide, SA , 5000'
      facialImageURL:
        type: string
        example: '/foo/images/img001.jpg'
      licenseNumber:
        type: string
        example: 'SAV-00001'
      criminalHistorySummary:
        type: string
        example: 'Very dangerous. Approach with caution.. Multiple aliases. Long criminal history'
      hasWarrantsOutstanding:
        type: string
        example: 'YES'
      hasFinesOutstanding:
        type: string
        example: 'NO'        
  PersonsOfInterestResponse:
    title: "Persons Of Interest Response"
    type: object
    properties:
      id:
        type: number
        default: 1
      fname:
        type: string
        default: 'Dave'
      lname:
        type: string
        default: 'Dundocerty'
      dob:
        type: string
        default: '19-MAR-2000'
      address:
        type: string
        default: '25 Grenfell Street, Adelaide, SA , 5000'
      facialImageURL:
        type: string
        default: '/foo/images/img001.jpg'
      licenseNumber:
        type: string
        default: 'SAV-00001'
      criminalHistorySummary:
        type: string
        default: 'Very dangerous. Approach with caution.. Multiple aliases. Long criminal history'
      hasWarrantsOutstanding:
        type: string
        default: 'YES'
      hasFinesOutstanding:
        type: string
        default: 'NO'   
...